requirements.txt
Django>=5.0.4,<6.0
djangorestframework>=3.15.1,<4.0
psycopg2>=2.9.9,<3.0
drf-spectacular>=0.27.2,<0.28
Pillow>=10.3.0,<10.4.0